grid07/requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0